Fix bugs and optimize prompt logs and formatting
