+++ examples/confild/conf/un_confild_case2.yaml
@@ -26,14 +26,14 @@ TRAIN:
 # U-Net Model Configuration (larger for Case2)
 UNET:
   image_size: 256                    # Latent space spatial size (larger for channel flow)
-  in_channels: 256                   # Input latent feature dimension
-  out_channels: 256                  # Output latent feature dimension
+  in_channels: 1                     # Input channels (single channel for latent codes treated as 2D images)
+  out_channels: 1                    # Output channels (single channel output)
   num_channels: 128                  # Base channel number for UNet
   num_res_blocks: 2                  # Number of residual blocks
   num_heads: 4                       # Number of attention heads
   num_head_channels: 64              # Channels per attention head
   attention_resolutions: "32,16,8"   # Attention resolutions
-  channel_mult: null                 # Channel multipliers (null for default)
+  channel_mult: [1, 2, 4, 8]         # Channel multipliers for each resolution level
 
 # Diffusion Process Configuration
 DIFF:
