+++ examples/confild/resample.py
@@ -46,10 +46,7 @@ class UniformSampler(ScheduleSampler):
 class LossAwareSampler(ScheduleSampler):
     def update_with_local_losses(self, local_ts, local_losses):
 
-        batch_sizes = [
-            th.to_tensor([0], dtype=th.int32, place=local_ts.device)
-            for _ in range(dist.get_world_size())
-        ]
+        batch_sizes = [th.to_tensor([0], dtype=th.int32, place=local_ts.device) for _ in range(dist.get_world_size())]
         dist.all_gather(
             batch_sizes,
             th.to_tensor([len(local_ts)], dtype=th.int32, place=local_ts.device),
@@ -63,9 +60,7 @@ class LossAwareSampler(ScheduleSampler):
         loss_batches = [th.zeros(max_bs).to(local_losses) for bs in batch_sizes]
         dist.all_gather(timestep_batches, local_ts)
         dist.all_gather(loss_batches, local_losses)
-        timesteps = [
-            x.item() for y, bs in zip(timestep_batches, batch_sizes) for x in y[:bs]
-        ]
+        timesteps = [x.item() for y, bs in zip(timestep_batches, batch_sizes) for x in y[:bs]]
         losses = [x.item() for y, bs in zip(loss_batches, batch_sizes) for x in y[:bs]]
         self.update_with_all_losses(timesteps, losses)
 
@@ -92,9 +87,7 @@ class LossSecondMomentResampler(LossAwareSampler):
         self.diffusion = diffusion
         self.history_per_term = history_per_term
         self.uniform_prob = uniform_prob
-        self._loss_history = np.zeros(
-            [diffusion.num_timesteps, history_per_term], dtype=np.float64
-        )
+        self._loss_history = np.zeros([diffusion.num_timesteps, history_per_term], dtype=np.float64)
         self._loss_counts = np.zeros([diffusion.num_timesteps], dtype=np.int32)
 
     def weights(self):
