+++ examples/symbolic_gn/conf/config_varogn.yaml
@@ -42,7 +42,7 @@ TRAIN:
     weight_decay: 1e-8
   lr_scheduler:
     name: "CosineAnnealingDecay"
-    gamma: 0.95  # 仅用于ExponentialDecay
+    gamma: 0.95  # Only used for ExponentialDecay
   loss:
     type: "MAE"
